+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 * Adding full log scale support to `xan plot`.
 * Adding `xan hist --scale`.
 * `xan window` is now able to run total aggregations.
+* Adding `thousands_sep`, `comma` and `significance` kwargs to `numfmt` moonblade function.
 
 *Fixes*